Move frequency (channel) utils into their own file
